remote_control/requirements.txt
discord.py
pytest
python-dotenv
